requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
